+++ lib/src/constants/colors.dart
@@ -6,4 +6,4 @@ Color success = const Color(0xFF61D800);
 Color normal = const Color(0xFF179DFF);
 Color warning = const Color(0xFFFF8B17);
 Color error = const Color(0xFFFF4D17);
-// Color defaultTextColor = const Color(0xFF707070);
+Color defaultTextColor = const Color(0xFF707070);