+++ lib/src/widgets/panara_confirm_dialog_widget.dart
@@ -7,7 +7,7 @@ import 'package:panara_dialogs/src/widgets/panara_button.dart';
 /// This is the PanaraConfirmDialogWidget.
 ///
 class PanaraConfirmDialogWidget extends StatelessWidget {
-  final String title;
+  final String? title;
   final String message;
   final String? imagePath;
   final String confirmButtonText;
@@ -20,7 +20,7 @@ class PanaraConfirmDialogWidget extends StatelessWidget {
   final Color? buttonTextColor;
   const PanaraConfirmDialogWidget({
     Key? key,
-    required this.title,
+    this.title,
     required this.message,
     required this.confirmButtonText,
     required this.cancelButtonText,
@@ -71,19 +71,21 @@ class PanaraConfirmDialogWidget extends StatelessWidget {
               const SizedBox(
                 height: 24,
               ),
-              Text(
-                title,
-                style: TextStyle(
-                  fontSize: 24,
-                  height: 1.2,
-                  fontWeight: FontWeight.w600,
-                  color: textColor,
+              if (title != null)
+                Text(
+                  title ?? "",
+                  style: TextStyle(
+                    fontSize: 24,
+                    height: 1.2,
+                    fontWeight: FontWeight.w600,
+                    color: textColor,
+                  ),
+                  textAlign: TextAlign.center,
                 ),
-                textAlign: TextAlign.center,
-              ),
-              const SizedBox(
-                height: 5,
-              ),
+              if (title != null)
+                const SizedBox(
+                  height: 5,
+                ),
               Text(
                 message,
                 style: TextStyle(